refactor: simplify getting started for newcomers (#160)
This commit combines fixes and improvements for the repository setup
process. It removes the self-signed SSL certificate generation, the
(default) mandatory HTTPS communication between the backend service and
the front-end, and other helper dependencies. Also, the changes are
ensuring that the setup finishes successfully on Linux systems as well.

For the ease of use, several new Make targets are added and the existing
targets are refactored. From now on, both the backend and front-end can
be managed through Make targets, from the repository root.

The default Mailhog Docker image is replaced by a fork that supports ARM
hosts, and smaller improvements are applied to eliminate the warnings,
raised by Docker.

To make bug reporting easier, a new bug report script was added, which
collects useful information from the host and backend service.

In addition, the CI steps are extended and improved to execute all tests
but k6.

Signed-off-by: Gábor Boros <gabor@opcotech.com>